+++ app/src/main/java/com/example/qraviaryapp/activities/dashboards/ExpensesActivity.kt
@@ -16,14 +16,12 @@ import android.view.View
 import android.widget.Button
 import android.widget.EditText
 import androidx.appcompat.app.AlertDialog
-import androidx.appcompat.app.AppCompatDelegate
 import androidx.core.content.ContextCompat
 import androidx.core.text.HtmlCompat
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.qraviaryapp.R
-import com.example.qraviaryapp.activities.AddActivities.AddBirdFlightActivity
 import com.example.qraviaryapp.activities.AddActivities.AddExpensesActivity
 import com.example.qraviaryapp.adapter.DetailedAdapter.ExpensesAdapter
 import com.google.firebase.auth.FirebaseAuth
@@ -116,7 +114,7 @@ class ExpensesActivity : AppCompatActivity() {
                     val dateValue = date.toString()
                     val commentValue = comment.toString()
                     data.expenses = mutationNameValue
-                    data.price = priceNameValue
+                    data.price = priceNameValue.toDouble()
                     data.expensesComment = commentValue
                     data.expensesDate = dateValue
                     if (Looper.myLooper() != Looper.getMainLooper()) {
@@ -170,7 +168,7 @@ class ExpensesActivity : AppCompatActivity() {
                 // Create a new ExpensesData object and add it to the dataList
                 val newExpensesData = ExpensesData()
                 newExpensesData.expenses = mutationNameValue
-                newExpensesData.price = PriceNameValue
+                newExpensesData.price = PriceNameValue.toDouble()
                 dataList.add(newExpensesData)
 
                 // Notify the adapter of the data change and sort the dataList
